requirements.txt
streamlit==1.23.1
openai==0.27.0
pandas==2.0.3
yfinance==0.2.27
plotly==5.10.0
python-dotenv==1.0.0
requests
pip>=24.2
setuptools>=67.0.0
wheel>=0.36.2
Pillow>=10.0.0
